+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#2781](https://github.com/open-telemetry/opentelemetry-python/pull/2781))
 - Fix tracing decorator with late configuration
   ([#2754](https://github.com/open-telemetry/opentelemetry-python/pull/2754))
+- Fix --insecure of CLI argument
+  ([#2696](https://github.com/open-telemetry/opentelemetry-python/pull/2696))
 
 ## [1.12.0rc2-0.32b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.12.0rc2) - 2022-07-04